change EGameState from enum to enum class

+++ source/sledge/sledgelib/netfuncs/game.cpp
@@ -5,7 +5,7 @@
 
 #define sledgelib_func extern "C" __declspec(dllexport)
 
-sledgelib_func unsigned int _GetState() { return g_Game->m_State; }
+sledgelib_func unsigned int _GetState() { return static_cast<unsigned int>(g_Game->m_State); }
 sledgelib_func void _SetState(unsigned int iState) { g_Game->m_NextState = static_cast<EGameState>(iState); }
 
 sledgelib_func bool IsPlaying() { return g_Game->m_Playing; }
